fix: disambiguate round() overload breaking Cardputer CI build
round(int) is ambiguous on the ESP32 3.2.5 toolchain (equally-ranked
float/double overloads). Pre-existing since the M5Unified refactor
(bb307d5); Cardputer builds have been failing CI since Jan 26.

+++ m5stick-nemo.ino
@@ -583,10 +583,10 @@ int screen_dim_current = 0;
 void screenBrightness(int bright){
   Serial.printf("Brightness: %d\n", bright);
   #if defined(AXP)
-    M5.Display.setBrightness(MINBRIGHT + round(((255 - MINBRIGHT) * bright / 100)));
+    M5.Display.setBrightness(MINBRIGHT + round((float)((255 - MINBRIGHT) * bright / 100)));
   #endif
   #if defined(BACKLIGHT)
-    int bl = MINBRIGHT + round(((255 - MINBRIGHT) * bright / 100)); 
+    int bl = MINBRIGHT + round((float)((255 - MINBRIGHT) * bright / 100));
     analogWrite(BACKLIGHT, bl);
   #endif
 }
